+++ pennylane/templates/layers/__init__.py
@@ -21,4 +21,5 @@ from .random import RandomLayers
 from .cv_neural_net import CVNeuralNetLayers
 from .simplified_two_design import SimplifiedTwoDesign
 from .basic_entangler import BasicEntanglerLayers
+from .particle_conserving_u2 import ParticleConservingU2
 from .particle_conserving_u1 import ParticleConservingU1